+++ microservices/account-service/src/main/java/com/mako/accountservice/repository/PasswordResetRepository.java
@@ -3,5 +3,7 @@ package com.mako.accountservice.repository;
 import com.mako.accountservice.model.PasswordReset;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PasswordResetRepository extends JpaRepository<PasswordReset, Long> {
+import java.math.BigInteger;
+
+public interface PasswordResetRepository extends JpaRepository<PasswordReset, BigInteger> {
 }
