+++ microservices/account-service/src/main/java/com/mako/accountservice/repository/UserRepository.java
@@ -3,6 +3,8 @@ package com.mako.accountservice.repository;
 import com.mako.accountservice.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<User, Long> {
+import java.math.BigInteger;
+
+public interface UserRepository extends JpaRepository<User, BigInteger> {
     User findByEmail(String email);
 }
