+++ microservices/account-service/src/main/java/com/mako/accountservice/service/UserService.java
@@ -4,14 +4,16 @@ import com.mako.accountservice.exception.UserNotFoundException;
 import com.mako.accountservice.model.PasswordReset;
 import com.mako.accountservice.model.User;
 
+import java.math.BigInteger;
+
 public interface UserService {
-    User getUser(Long id) throws UserNotFoundException;
+    User getUser(BigInteger id) throws UserNotFoundException;
 
     User save(User user);
 
     boolean isUserPresent(User user);
 
-    Long passwordResetRequest(PasswordReset data);
+    BigInteger passwordResetRequest(PasswordReset data);
 
     boolean passwordReset(PasswordReset data) throws UserNotFoundException;
